CustomerCoreService/requirements.txt
bcrypt==4.2.0
cryptography==43.0.1
fastapi[standard]==0.113.0
jinja2==3.1.5
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
requests==2.32.3
uvicorn==0.34.0